AI/requirement.txt
fastapi
uvicorn
youtube-transcript-api
langchain
langchain-core
langchain-cohere
langchain-groq
langchain-text-splitters
langchain-classic
faiss-cpu
deep-translator
python-dotenv
yt-dlp